+++ script/deployers/MumbaiMangroveFullTestnetDeployer.s.sol
@@ -93,14 +93,14 @@ contract MumbaiMangroveFullTestnetDeployer is Deployer {
     // Activate markets
     Tokens memory tokens;
     // AAVE tokens:
-    tokens.dai.token = IERC20(fork.get("DAI"));
-    tokens.crv.token = IERC20(fork.get("CRV"));
-    tokens.wbtc.token = IERC20(fork.get("WBTC"));
+    tokens.dai.token = IERC20(fork.get("DAI.T/AAVEv3"));
+    tokens.crv.token = IERC20(fork.get("CRV.T/AAVEv3"));
+    tokens.wbtc.token = IERC20(fork.get("WBTC.T/AAVEv3"));
     // Mangrove deployed tokens:
-    tokens.usdc.token = IERC20(fork.get("USDC"));
-    tokens.usdt.token = IERC20(fork.get("USDT"));
-    tokens.weth.token = IERC20(fork.get("WETH"));
-    tokens.wmatic.token = IERC20(fork.get("WMATIC"));
+    tokens.usdc.token = IERC20(fork.get("USDC.T/MGV"));
+    tokens.usdt.token = IERC20(fork.get("USDT.T/MGV"));
+    tokens.weth.token = IERC20(fork.get("WETH.T/MGV"));
+    tokens.wmatic.token = IERC20(fork.get("WMATIC.T/MGV"));
 
     // Get prices for AAVE tokens
     uint[] memory prices = priceOracle.getAssetsPrices(
@@ -154,13 +154,14 @@ contract MumbaiMangroveFullTestnetDeployer is Deployer {
     });
 
     // Deploy Kandel instance via KandelSeeder to get the Kandel contract verified
+    // Also write its address so it can be used as a library to call createGeometricDistribution
     new KandelSower().innerRun({
       kandelSeeder: seeder,
       olKeyBaseQuote: OLKey(address(markets[0].tkn1.token), address(markets[0].tkn2.token), markets[0].tickSpacing),
       sharing: false,
       onAave: false,
-      registerNameOnFork: false,
-      name: ""
+      registerNameOnFork: true,
+      name: "KandelLib"
     });
 
     // Deploy AaveKandel instance via AaveKandelSeeder to get the AaveKandel contract verified