+++ src/strategies/MangroveOrder.sol
@@ -36,12 +36,12 @@ contract MangroveOrder is Forwarder, IOrderLogic {
   ///@param factory the router proxy factory used to deploy or retrieve user routers
   ///@param deployer The address of the admin of `this` at the end of deployment
   constructor(IMangrove mgv, RouterProxyFactory factory, address deployer) Forwarder(mgv, factory, new SmartRouter()) {
-    setAdmin(deployer);
+    _setAdmin(deployer);
   }
 
   ///@inheritdoc IOrderLogic
   ///@dev We also allow Mangrove to call this so that it can part of an offer logic.
-  function setExpiry(bytes32 olKeyHash, uint offerId, uint date) public mgvOrOwner(olKeyHash, offerId) {
+  function setExpiry(bytes32 olKeyHash, uint offerId, uint date) public onlyOwner(olKeyHash, offerId) {
     expiring[olKeyHash][offerId] = date;
     emit SetExpiry(olKeyHash, offerId, date);
   }
@@ -79,7 +79,7 @@ contract MangroveOrder is Forwarder, IOrderLogic {
   ///@dev Calling this function, with the `deprovision` flag, on an offer that is already retracted must be used to retrieve the locked provisions.
   function retractOffer(OLKey memory olKey, uint offerId, bool deprovision)
     public
-    mgvOrOwner(olKey.hash(), offerId)
+    onlyOwner(olKey.hash(), offerId)
     returns (uint freeWei)
   {
     return _retractOffer(olKey, offerId, deprovision);