+++ test/strategies/unit/AccessControl.t.sol
@@ -2,55 +2,34 @@
 pragma solidity ^0.8.10;
 
 import {StratTest} from "@mgv-strats/test/lib/StratTest.sol";
-
-import {DirectTester} from "@mgv-strats/src/toy_strategies/offer_maker/DirectTester.sol";
-import {AbstractRouter, RL} from "@mgv-strats/src/strategies/routers/abstract/AbstractRouter.sol";
-import {TestToken} from "@mgv/test/lib/tokens/TestToken.sol";
-import {IMangrove} from "@mgv/src/IMangrove.sol";
-import {IERC20} from "@mgv/lib/IERC20.sol";
+import {AccessControlled} from "@mgv-strats/src/strategies/utils/AccessControlled.sol";
 
 contract AccessControlTest is StratTest {
-  TestToken weth;
-  TestToken usdc;
+  event SetAdmin(address admin);
+
   address payable admin;
-  DirectTester makerContract;
+  AccessControlled internal testContract;
 
   function setUp() public virtual override {
-    options.base.symbol = "WETH";
-    options.quote.symbol = "USDC";
-    options.quote.decimals = 6;
-    options.defaultFee = 30;
-
     super.setUp();
-    // rename for convenience
-    weth = base;
-    usdc = quote;
-
     admin = freshAddress("admin");
-    deal(admin, 1 ether);
 
-    DirectTester.RouterParams memory noRouter;
-    vm.prank(admin);
-    makerContract = new DirectTester({
-      mgv: IMangrove($(mgv)),
-      routerParams: noRouter
-    });
-
-    vm.startPrank(admin);
-    weth.approve(address(makerContract), type(uint).max);
-    usdc.approve(address(makerContract), type(uint).max);
-    vm.stopPrank();
+    vm.expectEmit(true, true, true, true);
+    emit SetAdmin(admin);
+    testContract = new AccessControlled(admin);
   }
 
   function testCannot_setAdmin() public {
     vm.expectRevert("AccessControlled/Invalid");
-    makerContract.setAdmin(freshAddress());
+    testContract.setAdmin(freshAddress());
   }
 
   function test_admin_can_set_admin() public {
     address newAdmin = freshAddress("newAdmin");
+    expectFrom($(testContract));
+    emit SetAdmin(newAdmin);
     vm.prank(admin);
-    makerContract.setAdmin(newAdmin);
-    assertEq(makerContract.admin(), newAdmin, "Incorrect admin");
+    testContract.setAdmin(newAdmin);
+    assertEq(testContract.admin(), newAdmin, "Incorrect admin");
   }
 }